requirements.txt
openai
python-dotenv
requests
google-api-python-client
google-auth
google-auth-oauthlib